TollCalculator: Moves concrete Vehicle:s to separate package
